Now separating ANT Runtime APIs from ant.js
