+++ lite/api-modules/ant/js/ant.js
@@ -4,85 +4,13 @@ var http = require('http');
 var RESULT_SUCCESS = 'Success';
 var RESULT_FAILED = 'Failed';
 
-/** Runtime API start **/
-var RuntimeAPI = { _current_app: undefined };
-
-RuntimeAPI._removeCurrentApp = function () {
-  this._current_app = undefined;
-};
-
-RuntimeAPI.setCurrentApp = function (onInitialize, onStart, onStop) {
-  if (!(onInitialize instanceof Function) || !(onStart instanceof Function) ||
-    !(onStop instanceof Function)) {
-    return RESULT_FAILED;
-  }
-  this._current_app = new App(onInitialize, onStart, onStop);
-
-  return RESULT_SUCCESS;
-};
-
-RuntimeAPI.getCurrentApp = function () {
-  return this._current_app;
-};
-
-/* App start */
-function App(onInitialize, onStart, onStop) {
-  this.state = this.STATE.IDLE;
-  this.onInitialize = onInitialize;
-  this.onStart = onStart;
-  this.onStop = onStop;
-
-  this.onInitialize();
-}
-
-App.prototype.STATE = {};
-App.prototype.STATE.IDLE = 0;
-App.prototype.STATE.RUNNING = 1;
-
-App.prototype.start = function () {
-  if (this.state == this.STATE.RUNNING) {
-    return RESULT_FAILED;
-  }
-  this.state = this.STATE.RUNNING;
-  this.onStart();
-  return RESULT_SUCCESS;
-};
-
-App.prototype.stop = function () {
-  if (this.state == this.STATE.IDLE) {
-    return RESULT_FAILED;
-  }
-  this.state = this.STATE.IDLE;
-  this.onStop();
-  return RESULT_SUCCESS;
-};
-
-App.prototype.getState = function () {
-  if (this.state == this.STATE.IDLE) {
-    return 'Idle';
-  } else if (this.state == this.STATE.RUNNING) {
-    return 'Running';
-  } else {
-    return 'Unknown';
-  }
-};
-
-App.prototype.getInfo = function () {
-  var appInfo = { 'state': this.getState() };
-  return appInfo;
-};
-/* App end */
-
-/** Runtime API end **/
-
-
 /** Stream API start **/
 // TODO: native-side Stream API implementation is required.
 function StreamAPI() { }
 StreamAPI.createPipeline = function (pipeline_name) {
   return new Pipeline(pipeline_name);
 };
-StreamAPI.createElement() = function (element_name) {
+StreamAPI.createElement = function (element_name) {
   return new Element(element_name);
 };
 
@@ -145,6 +73,7 @@ CompanionAPI.sendMessage = function (message) {
     return false;
   }
 
+  // IoT.js ----
   var options = {
     method: 'POST',
     host: this._companionHost,
@@ -155,6 +84,14 @@ CompanionAPI.sendMessage = function (message) {
   var client_request = http.request(options);
   client_request.write(message);
   client_request.end();
+  // node.js ----
+  // var options = {
+  //   method: 'POST',
+  //   uri: 'http://' + this._companionHost + ':' + this._companionPort + '/' + this._companionPath,
+  //   body: message
+  // };
+  // request.post(options, function (err, httpResponse, body) {
+  // });
   return true;
 };
 
@@ -185,7 +122,7 @@ CompressionServerAPI.downloadKernel = function (kernelId) {
 /** ANT main object start **/
 function ANT() { }
 
-ANT.prototype.runtime = RuntimeAPI;
+ANT.prototype.runtime = require('ant_runtime');
 ANT.prototype.stream = StreamAPI;
 ANT.prototype.companion = CompanionAPI;
 ANT.prototype.compressionServer = CompressionServerAPI;
