+++ lite/api-modules/ant/modules.json
@@ -4,7 +4,9 @@
       "js_file": "js/ant.js",
       "require": [
         "buffer",
-        "console"
+        "console",
+        "ant_runtime",
+        "http"
       ]
     }
   }
