+++ lite/test/build-iotjs-module-on-rpi3.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 MODULE_PATH=../../lite/api-modules/
 cd ../../dep/iotjs/
-./tools/build.py --target-board=rpi3 --cmake-param=-DENABLE_MODULE_ANT=ON \
-    --external-modules=${MODULE_PATH}/ant
+./tools/build.py --target-board=rpi3 --cmake-param='-DENABLE_MODULE_ANT=ON -DENABLE_MODULE_ANT_RUNTIME=ON' \
+    --external-modules=${MODULE_PATH}/ant,${MODULE_PATH}/ant_runtime
 cd -
 
 mkdir -p iotjs
