+++ lite/test/build-nodejs-module.sh
@@ -4,5 +4,7 @@ cd nodejs
 
 MODULE_PATH=../../api-modules/
 npm install ${MODULE_PATH}/ant
+# request: for HTTP request
+npm install request
 
 cp -r ../../app-runtime/* ./
